save order

+++ src/main/java/com/efreh/order_manager/dao/OrderRepository.java
@@ -3,5 +3,6 @@ package com.efreh.order_manager.dao;
 import com.efreh.order_manager.entity.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order,Integer> {
+public interface OrderRepository extends JpaRepository<Order,Long> {
+
 }
